Remove queries.json, move config to services.json
Having two files for services is mistake prone. Similar to database
schemas, it's useful to put all services related data in the same file.

This commit moves the content of queries.json into services.json, and
refactors linkchanbot to handle that. This means one less global
variable, and less configuration logic.

This should not be breaking but a minor inconvenience. Simply remove
queries.json and remove (or backup) services.json, than reinstall.