Ignore messaged forwarded from self
Messages from the bot which have been forwarded to group chats are
ignored. This avoids duplicate substitution messages.

Messages from the bot which have been forwarded to the bot as a private
message are not ignored. This continues to allow for intentional
duplication or replacement of old faulty proxy links for newer ones.